moved midinotes into grid component
